Created helpers, extracted from applicationhelper
