+++ app/helpers/select_helper.rb
@@ -1,0 +1,42 @@
+module SelectHelper
+
+  def day_choices
+    [
+      ["Last Day", "1"],
+      ["Last Seven Days", "7"],
+      ["Last Thirty Days", "30"]
+    ]
+  end
+
+  def section_choices
+    [
+      ["Arts", "arts"],
+      ["Automobiles", "automobiles"],
+      ["Books", "books"],
+      ["Business", "business"],
+      ["Fashion", "fashion"],
+      ["Food", "food"],
+      ["Health", "health"],
+      ["Home", "home"],
+      ["Insider", "insider"],
+      ["Magazine", "magazine"],
+      ["Movies", "movies"],
+      ["National", "national"],
+      ["NY Region", "nyregion"],
+      ["Obituaries", "obituaries"],
+      ["Opinion", "opinion"],
+      ["Politics", "politics"],
+      ["Real Estate", "realestate"],
+      ["Science", "science"],
+      ["Sports", "sports"],
+      ["Sunday Review", "sundayreview"],
+      ["Technology", "technology"],
+      ["Theater", "theater"],
+      ["T Magazine", "tmagazine"],
+      ["Travel", "travel"],
+      ["Upshot", "upshot"],
+      ["World", "world"]
+    ]
+  end
+
+end
